requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.0.0
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
utils==1.0.1
(venv) redihoxha@MacBook-Pro banking_project1 % pip freeze
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.0.0
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
utils==1.0.1
